Refactor Link to use ASWebAuthenticationSession (#2643)
