--- Stripe/StripeiOSTests/LinkNoticeViewSnapshotTests.swift
@@ -1,51 +1,0 @@
-//
-//  LinkNoticeViewSnapshotTests.swift
-//  StripeiOS Tests
-//
-//  Created by Ramon Torres on 3/31/22.
-//  Copyright © 2022 Stripe, Inc. All rights reserved.
-//
-
-import iOSSnapshotTestCase
-import UIKit
-
-@testable@_spi(STP) import Stripe
-@testable@_spi(STP) import StripeCore
-@testable@_spi(STP) import StripePayments
-@testable@_spi(STP) import StripePaymentSheet
-@testable@_spi(STP) import StripePaymentsUI
-
-class LinkNoticeViewSnapshotTests: FBSnapshotTestCase {
-
-    override func setUp() {
-        super.setUp()
-        //        recordMode = true
-    }
-
-    func testError() {
-        verify(
-            LinkNoticeView(
-                type: .error,
-                text:
-                    "This card has expired. Update it to keep using it or use a different payment method."
-            )
-        )
-    }
-
-    func verify(
-        _ sut: UIView,
-        identifier: String? = nil,
-        file: StaticString = #filePath,
-        line: UInt = #line
-    ) {
-        let size = sut.systemLayoutSizeFitting(
-            CGSize(width: 340, height: UIView.layoutFittingCompressedSize.height),
-            withHorizontalFittingPriority: .required,
-            verticalFittingPriority: .fittingSizeLevel
-        )
-
-        sut.bounds = CGRect(origin: .zero, size: size)
-        STPSnapshotVerifyView(sut, identifier: identifier, file: file, line: line)
-    }
-
-}
