+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFlowController.swift
@@ -254,22 +254,7 @@ extension PaymentSheet {
                 self.isPresented = true
             }
 
-            if let linkAccount = LinkAccountContext.shared.account,
-               linkAccount.sessionState == .requiresVerification,
-               !linkAccount.hasStartedSMSVerification {
-                let verificationController = LinkVerificationController(linkAccount: linkAccount)
-                verificationController.present(from: presentingViewController) { [weak self] result in
-                    switch result {
-                    case .completed:
-                        self?.viewController.selectLink()
-                        completion?()
-                    case .canceled, .failed:
-                        showPaymentOptions()
-                    }
-                }
-            } else {
-                showPaymentOptions()
-            }
+            showPaymentOptions()
         }
 
         /// Completes the payment or setup.
