+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentSheetFlowControllerViewController.swift
@@ -519,24 +519,7 @@ extension PaymentSheetFlowControllerViewController: SavedPaymentOptionsViewContr
 extension PaymentSheetFlowControllerViewController: AddPaymentMethodViewControllerDelegate {
     func didUpdate(_ viewController: AddPaymentMethodViewController) {
         error = nil  // clear error
-
-        if case .link(let linkOption) = selectedPaymentOption,
-           let linkAccount = linkOption.account,
-           linkAccount.sessionState == .requiresVerification {
-            isVerificationInProgress = true
-            updateUI()
-
-            let verificationController = LinkVerificationController(mode: .inlineLogin, linkAccount: linkAccount)
-            verificationController.present(from: self, completion: { [weak self] _ in
-                // Verification result is ignored here on purpose. If verification is canceled or fails,
-                // we will simply don't block the payment. This will be revised after we redesign the inline
-                // signup UI to include verification status.
-                self?.isVerificationInProgress = false
-                self?.updateUI()
-            })
-        } else {
-            updateUI()
-        }
+        updateUI()
     }
 
     func shouldOfferLinkSignup(_ viewController: AddPaymentMethodViewController) -> Bool {
